Revert "MyTodoApp implemented in FLUX"
