+++ src/js/components/Footer.js
