+++ src/js/components/TodoList/Item.js
@@ -3,21 +3,20 @@ import React from "react";
 export default class Item extends React.Component {
     // handles the change (when the list item status is changed) [checkbox is
     // checked or unchecked]
-    // sends the index number (id) of the list to updateListItemStatus in
-    // to update state layout.js
+    // sends the index number of the list to updateListItemStatus in layout.js
+    // to update state
     handleChange(event) {
         this.props.updateListItemStatus(event);
     }
 
-    // handles the click on close button, sends the index no (id) to the
-    // function in layout.js removeListItem function
+    // handles the click on close button, sends the index no to the function in
+    // layout.js removeListItem function
     handleClick(event) {
         this.props.removeListItem(event);
     }
 
     render() {
-        const text = this.props.item.text;
-        const done = this.props.item.done;
+        const { text, done } = this.props.item;
         const index = this.props.index;
 
         return (
