--- src/js/dispatcher.js
@@ -1,5 +1,0 @@
-// Dispatcher to receive actions (TodoActions) from components and forwards
-// them to store (TodoStore)
-import { Dispatcher } from "flux";
-
-export default new Dispatcher;
